adds auto incremental id, adds the second table to try this

+++ src/main/java/com/MongoSpring/MongoSpring/Repository/StudentRepo.java
@@ -3,5 +3,6 @@ package com.MongoSpring.MongoSpring.Repository;
 import com.MongoSpring.MongoSpring.Model.Student;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface StudentRepo extends MongoRepository<Student, Integer> {
+public interface StudentRepo extends MongoRepository<Student, Long> {
+
 }
